+++ src/main/java/com/train/auction/action/MgrItemAction.java
@@ -1,11 +1,11 @@
 package com.train.auction.action;
 
-import java.util.*;
+import java.math.BigInteger;
+import java.util.List;
+import java.util.Map;
 
-import com.opensymphony.xwork2.*;
+import com.opensymphony.xwork2.ActionContext;
 import com.train.auction.action.base.BaseActionInterface;
-import com.train.auction.exception.AuctionException;
-import com.train.auction.service.AuctionManager;
 
 
 public class MgrItemAction extends BaseActionInterface
@@ -16,7 +16,7 @@ public class MgrItemAction extends BaseActionInterface
 	public String execute() throws Exception
 	{
 		Map session = ActionContext.getContext().getSession();
-		Integer userId = (Integer)session.get("userId");
+		BigInteger userId = (BigInteger)session.get("userId");
 		setItems(mgr.getItemsByOwner(userId));
 		setKinds(mgr.getAllKind());
 		return SUCCESS;
