+++ src/main/java/com/train/auction/action/ViewBidAction.java
@@ -1,11 +1,11 @@
 package com.train.auction.action;
 
-import java.util.*;
+import java.math.BigInteger;
+import java.util.List;
+import java.util.Map;
 
 import com.opensymphony.xwork2.ActionContext;
 import com.train.auction.action.base.BaseAction;
-import com.train.auction.exception.AuctionException;
-import com.train.auction.service.AuctionManager;
 
 public class ViewBidAction extends BaseAction
 {
@@ -14,7 +14,7 @@ public class ViewBidAction extends BaseAction
 	public String execute()throws Exception
 	{
 		Map session = ActionContext.getContext().getSession();
-		Integer userId = (Integer)session.get("userId");
+		BigInteger userId = (BigInteger)session.get("userId");
 		setBids(mgr.getBidByUser(userId));
 		return SUCCESS;
 	}
